+++ app/src/main/admin/components/Touchstones/List/TouchstoneList.tsx
@@ -1,6 +1,5 @@
 import {Touchstone} from "../../../../shared/models/Generated";
 import * as React from "react";
-import {groupBy, isNullOrEmptyArray} from "../../../../shared/Helpers";
 import {AdminAppState} from "../../../reducers/adminAppReducers";
 import {compose} from "recompose";
 import {connect} from "react-redux";
@@ -19,24 +18,14 @@ export const TouchstoneListComponent: React.SFC<TouchstoneListProps> = (props: T
         </tr>
         </thead>
         <tbody>
-        {props.touchstones.map(t => <TouchstoneListItem key={t.name} {...t} />)}
+        {props.touchstones.map(t => <TouchstoneListItem key={t.id} {...t} />)}
         </tbody>
     </table>;
 };
 
-function getTouchstonesFromVersions(touchstones: Touchstone[]): Touchstone[] {
-    if (isNullOrEmptyArray(touchstones)) {
-        return [];
-    }
-    return groupBy(touchstones, t => t.name).map(g => {
-        g.values.sort(t => t.version);
-        return g.values[0];
-    });
-}
-
 export const mapStateToProps = (state: AdminAppState): TouchstoneListProps => {
     return {
-        touchstones: getTouchstonesFromVersions(state.touchstones.touchstones)
+        touchstones: state.touchstones.touchstones
     }
 };
