+++ app/src/main/shared/Helpers.ts
@@ -76,26 +76,3 @@ export const helpers = {
         return "?redirectUrl=" + encodeURI(settings.montaguUrl() + redirectPath);
     }
 };
-};
-export const isNullOrEmptyArray = (arrayData: any[]) => {
-    return !Array.isArray(arrayData) || !arrayData.length;
-};
-
-interface Grouping<TKey, TValue> {
-    key: TKey,
-    values: TValue[]
-}
-
-export function groupBy<TItem>(
-    items: TItem[],
-    groupingFunction: (item: TItem) => string
-): Grouping<string, TItem>[] {
-    const result: { [index: string]: Grouping<string, TItem> } = {};
-    items.forEach(item => {
-        const key = groupingFunction(item);
-        if (!result[key]) {
-            result[key] = {key, values: []};
-        }
-        result[key].values.push(item);
-    });
-    return Object.keys(result).map(x => result[x]);
